Change copy_constructor C++ code

+++ various/pointer_to_pointer_allocation/pointer_to_pointer_allocation.c
@@ -24,7 +24,7 @@ int main(){
 	srand(time(NULL));
 	for(int i = 0; i < size;++i){
 		for(int j = 0; j < size;++j){
-			mat[i][j] = rand() * rand_limit / RAND_MAX + 1;
+			mat[i][j] = rand() * rand_limit / (float) RAND_MAX + 1;
 		}
 	}
